extracted test model to a dedicated project which is now used by different integration tests - for SQL and noSQL databases. Adopted test implementations to the changes in the model

+++ test-model/src/main/java/eu/drus/jpa/unit/test/model/InstantAccessAccount.java
@@ -25,8 +25,8 @@ public class InstantAccessAccount extends Account {
 
     @Override
     public float transfer(final float amount, final Account toAccount) {
-        final float balance = getBalance();
-        final float balanceAfterTransfer = balance - amount;
+        final double balance = getBalance();
+        final double balanceAfterTransfer = balance - amount;
         if (balanceAfterTransfer < 0.0f) {
             return 0;
         }
